school/requirements.txt
asgiref==3.7.2
certifi==2022.12.7
charset-normalizer==3.3.2
distlib==0.3.6
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
Pillow==10.1.0
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2==2.9.9
pytz==2023.3.post1
reportlab==4.0.7
requests==2.31.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.1.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
